fix create_data() int problem

+++ test_vpython_4.py
@@ -64,7 +64,7 @@ class rabo():
             create_data.append(0)
         create_data.pop()
         for i in range(1, len(create_data)-1, 2):
-            create_data[i] = int((create_data[i+1] + create_data[i-1]) / 2) 
+            create_data[i] = (create_data[i+1] + create_data[i-1]) / 2 
 
     def the_canvas(self):
         self.scene = canvas(title='1', width=800, height=800, x=0, y=0, center=vector(0,0.06,0), background=vector(0.5,0.6,0.5))
